See RN5.3.0 (20260430)
- Migrate all Jingle stanza sid to use UUIDv4 as recommended in XEP-0353: Jingle Message Initiation 0.8.0 (2026-02-19)
  i.e. 5. Business Rules https://xmpp.org/extensions/xep-0353.html#business-rules
- Update JingleMessage design/implementation to compliance with v0.8.0 (2026-02-19) specifications.
- During tie-break, sending of Retract and Reject message to remote Jid will not follow as specified; else JM does not work properly.
- Ensure proper handling of jingle call when exceptions occurred; JingleMessage session needs to be informed and hang up call correctly.
- aTalk media call now handles Tie-Break, Call-Migrate, Call-Waiting, Call-Hold and Call-Transfer.
- aTalk JingleMessage implementation is backward compatible to client with XEP-0353 v0.3.1 implementation e.g. conversions.
- Fix Call Switch/Transfer creating multiple instances of VideoCallActivity when access via notification; Use CallManager finish(), and then startActivity.
  When switching between two calls from notification, a new instance of the VideoCallActivity is created each time.
  Need user to hangup each newly created activity when call end.
  Tested no working: a. ActivityManager.moveTaskToFront();  b. FLAG_ACTIVITY_REORDER_TO_FRONT.
- Remove CallPeerJabberImpl#hangup() method never throws exceptions.
- Fix presence indicator error even after the user has authenticated. Roster callback on onRosterLoaded() is not called due to race condition.
  Move addRosterLoadedListener() after user is connected and not after registered.
- Add RegistrationState.CONNECTED; all discovered features must be added before user is authenticated.
- Change incorrect use 'delete' to 'retract' to remove a node item in onPepEvent received omemo#devices itemId != 'current'.
- Update smack pubsub RetractItem implementation to confirm to 1.30.0 (2025-11-11)
- Update 'About' to reflect actual third party libraries used, and use atalk.doap instead for XEP used by aTalk.
- Upgrade jni libs for libvpx 1.15.2 => 1.16.0; opus lib 1.5.2 => 1.6.1; libOgg 1.3.5 => 1.3.6; update init_libSppexOgg.sh script file.

+++ aTalk/jni/speex/init_libSpeexOgg.sh
@@ -26,40 +26,42 @@ fi
 if [[ $# -eq 2 ]]; then
   LIB_OGG_GIT=$2
 else
-  LIB_OGG_GIT="v1.3.5"
+  LIB_OGG_GIT="libogg-1.3.6"
 fi
 echo -e "\n### Fetch (speex-1.2rc1 & ${LIB_SPEEX_GIT}) and (ogg ${LIB_OGG_GIT}) libraries source ###"
 
 LIB_SPEEX="libspeex"
+proceed=true
 if [[ -d ${LIB_SPEEX} ]]; then
   # version="$(grep '^#define SPEEX_VERSION' < ${LIB_SPEEX}/arch.h | sed 's/^.*\([1-9]\.[0-9][a-z]*[0-9]\).*$/\1/')"
   version="$(grep '^#define SPEEX_VERSION' < ${LIB_SPEEX}/arch.h | sed 's/^.*\([1-9]\.[0-9]\.[0-9]\).*$/\1/')"
   if [[ "${LIB_SPEEX_GIT}" =~ .*"${version}".* ]]; then
     echo -e "========== Current speex source is: ${LIB_SPEEX}-${version} ==========\n"
+    proceed=false
+  fi
+fi
+
+if $proceed; then
+  rm -rf ${LIB_SPEEX}
+  echo -e "\n================ Fetching library source for ${LIB_SPEEX}: ${LIB_SPEEX_GIT} ============================"
+  wget -O- https://ftp.osuosl.org/pub/xiph/releases/speex/speex-1.2rc1.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_SPEEX}
+  wget -O- https://ftp.osuosl.org/pub/xiph/releases/speex/${LIB_SPEEX_GIT}.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_SPEEX}
+  echo -e "======== Completed speex library source update ============================"
+fi
+
+## ==================== libogg ================= ##
+LIB_OGG="libogg"
+if [[ -d ${LIB_OGG} ]]; then
+  version="$(grep -m 1 '^Version' < ${LIB_OGG}/CHANGES | sed 's/^Version \([1-9]\.[0-9]\.[0-9]\).*$/libogg-\1/')"
+  if [[ "${LIB_OGG_GIT}" =~ .*"${version}".* ]]; then
+    echo -e "\n========== Current ogg source is: ${LIB_OGG}-${version} =========="
     exit 0
   fi
 fi
 
-rm -rf ${LIB_SPEEX}
-echo -e "\n================ Fetching library source for ${LIB_SPEEX}: ${LIB_SPEEX_GIT} ============================"
-wget -O- http://downloads.us.xiph.org/releases/speex/speex-1.2rc1.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_SPEEX}
-wget -O- http://downloads.us.xiph.org/releases/speex/${LIB_SPEEX_GIT}.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_SPEEX}
-echo -e "======== Completed speex library source update ============================"
-
-
-## ==================== libogg ================= ##
-LIB_OGG="libogg"
-#if [[ -d ${LIB_OGG} ]]; then
-#  version="$(grep '^PACKAGE_VERSION' < ${LIB_OGG}/package_version | sed 's/^.*\([1-9]\.[0-9]\.[0-9]\).*$/\1/')"
-#  if [[ "${LIB_OGG_GIT}" =~ .*"${version}".* ]]; then
-#    echo -e "\n========== Current ogg source is: ${LIB_OGG}-${version} =========="
-#    exit 0
-#  fi
-#fi
-
 rm -rf ${LIB_OGG}
 echo -e "\n================ Fetching library source for ${LIB_OGG}: ${LIB_OGG_GIT} ============================"
-wget -O- https://github.com/xiph/ogg/archive/refs/tags/${LIB_OGG_GIT}.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_OGG}
+wget -O- https://ftp.osuosl.org/pub/xiph/releases/ogg/${LIB_OGG_GIT}.tar.gz | tar xz --strip-components=1 --one-top-level=${LIB_OGG}
 echo -e "======== Completed ogg library source update ============================"
 
 # For testing only
